requirements.txt
streamlit
tensorflow-cpu
pillow
rembg[cpu]
numpy